add: tipping page step 1
